Вынесение общих полей документов в DocumentEntity
