+++ src/main/java/com/lognex/api/entities/documents/CommissionReportInDocumentEntity.java
@@ -2,7 +2,6 @@ package com.lognex.api.entities.documents;
 
 import com.lognex.api.entities.*;
 import com.lognex.api.entities.agents.AgentEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.entities.agents.OrganizationEntity;
 import com.lognex.api.responses.ListEntity;
 import lombok.EqualsAndHashCode;
@@ -17,7 +16,6 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class CommissionReportInDocumentEntity extends DocumentEntity {
-    private String accountId;
     private AgentEntity agent;
     private Boolean applicable;
     private LocalDateTime commissionPeriodEnd;
@@ -27,23 +25,16 @@ public class CommissionReportInDocumentEntity extends DocumentEntity {
     private String created;
     private MetaEntity documents;
     private String externalCode;
-    private GroupEntity group;
-    private String id;
     private LocalDateTime moment;
-    private String name;
     private OrganizationEntity organization;
-    private EmployeeEntity owner;
     private Integer payedSum;
     private ListEntity positions;
     private RateEntity rate;
     private Double rewardPercent;
     private RewardType rewardType;
-    private Boolean shared;
     private StateEntity state;
     private Long sum;
-    private LocalDateTime updated;
     private Boolean vatEnabled;
     private Boolean vatIncluded;
     private Long vatSum;
-    private Integer version;
 }
