+++ src/main/java/com/lognex/api/entities/documents/CustomerOrderDocumentEntity.java
@@ -1,8 +1,10 @@
 package com.lognex.api.entities.documents;
 
-import com.lognex.api.entities.*;
+import com.lognex.api.entities.ContractEntity;
+import com.lognex.api.entities.MetaEntity;
+import com.lognex.api.entities.RateEntity;
+import com.lognex.api.entities.StateEntity;
 import com.lognex.api.entities.agents.AgentEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.entities.agents.OrganizationEntity;
 import com.lognex.api.responses.ListEntity;
 import lombok.EqualsAndHashCode;
@@ -21,7 +23,6 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class CustomerOrderDocumentEntity extends DocumentEntity {
-    private String accountId;
     private AgentEntity agent;
     private Boolean applicable;
     private ContractEntity contract;
@@ -30,26 +31,19 @@ public class CustomerOrderDocumentEntity extends DocumentEntity {
     private String description;
     private MetaEntity documents;
     private String externalCode;
-    private GroupEntity group;
-    private String id;
     private Long invoicedSum;
     private LocalDateTime moment;
-    private String name;
     private OrganizationEntity organization;
-    private EmployeeEntity owner;
     private Long payedSum;
     private ListEntity positions;
     private RateEntity rate;
     private Long reservedSum;
-    private Boolean shared;
     private Long shippedSum;
     private StateEntity state;
     private MetaEntity store;
     private Long sum;
     private String syncId;
-    private LocalDateTime updated;
     private Boolean vatEnabled;
     private Boolean vatIncluded;
     private Long vatSum;
-    private Integer version;
 }
