+++ src/main/java/com/lognex/api/entities/documents/DocumentEntity.java
@@ -1,10 +1,14 @@
 package com.lognex.api.entities.documents;
 
+import com.lognex.api.entities.GroupEntity;
 import com.lognex.api.entities.MetaEntity;
+import com.lognex.api.entities.agents.EmployeeEntity;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.time.LocalDateTime;
 
 /**
  * Сущность, имеющая поле Метаданных
@@ -15,6 +19,15 @@ import lombok.Setter;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public abstract class DocumentEntity extends MetaEntity {
+    private String id;
+    private GroupEntity group;
+    private Integer version;
+    private String accountId;
+    private String name;
+    private LocalDateTime updated;
+    private Boolean shared;
+    private EmployeeEntity owner;
+
     @Getter
     @Setter
     @NoArgsConstructor
