+++ src/main/java/com/lognex/api/entities/documents/InventoryDocumentEntity.java
@@ -1,9 +1,7 @@
 package com.lognex.api.entities.documents;
 
-import com.lognex.api.entities.GroupEntity;
 import com.lognex.api.entities.MetaEntity;
 import com.lognex.api.entities.StoreEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.entities.agents.OrganizationEntity;
 import com.lognex.api.responses.ListEntity;
 import lombok.EqualsAndHashCode;
@@ -18,19 +16,11 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class InventoryDocumentEntity extends DocumentEntity {
-    private String accountId;
     private MetaEntity documents;
     private String externalCode;
-    private GroupEntity group;
-    private String id;
     private LocalDateTime moment;
-    private String name;
     private OrganizationEntity organization;
-    private EmployeeEntity owner;
     private ListEntity positions;
-    private Boolean shared;
     private StoreEntity store;
     private Long sum;
-    private LocalDateTime updated;
-    private Integer version;
 }
