+++ src/main/java/com/lognex/api/entities/documents/ProcessingOrderDocumentEntity.java
@@ -1,11 +1,9 @@
 package com.lognex.api.entities.documents;
 
-import com.lognex.api.entities.GroupEntity;
 import com.lognex.api.entities.MetaEntity;
 import com.lognex.api.entities.StateEntity;
 import com.lognex.api.entities.StoreEntity;
 import com.lognex.api.entities.agents.AgentEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.responses.ListEntity;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -19,23 +17,15 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class ProcessingOrderDocumentEntity extends DocumentEntity {
-    private String accountId;
     private Boolean applicable;
     private String description;
     private MetaEntity documents;
     private String externalCode;
-    private GroupEntity group;
-    private String id;
     private LocalDateTime moment;
-    private String name;
     private AgentEntity organization;
-    private EmployeeEntity owner;
     private ListEntity positions;
     private ProcessingPlanDocumentEntity processingPlan;
     private Double quantity;
-    private Boolean shared;
     private StateEntity state;
     private StoreEntity store;
-    private LocalDateTime updated;
-    private Integer version;
 }
