+++ src/main/java/com/lognex/api/entities/documents/ProcessingPlanDocumentEntity.java
@@ -1,8 +1,6 @@
 package com.lognex.api.entities.documents;
 
-import com.lognex.api.entities.GroupEntity;
 import com.lognex.api.entities.MetaEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.entities.products.markers.ProductMarker;
 import com.lognex.api.responses.ListEntity;
 import lombok.EqualsAndHashCode;
@@ -10,26 +8,16 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class ProcessingPlanDocumentEntity extends DocumentEntity {
-    private String accountId;
     private Long cost;
     private String externalCode;
-    private GroupEntity group;
-    private String id;
     private ListEntity materials;
-    private String name;
-    private EmployeeEntity owner;
     private String pathName;
     private ListEntity products;
-    private Boolean shared;
-    private LocalDateTime updated;
-    private Integer version;
 
     @Getter
     @Setter
