+++ src/main/java/com/lognex/api/entities/documents/RetailDrawerCashOutEntity.java
@@ -1,9 +1,10 @@
-package com.lognex.api.entities;
+package com.lognex.api.entities.documents;
 
+import com.lognex.api.entities.MetaEntity;
+import com.lognex.api.entities.RateEntity;
+import com.lognex.api.entities.RetailShiftEntity;
 import com.lognex.api.entities.agents.AgentEntity;
-import com.lognex.api.entities.agents.EmployeeEntity;
 import com.lognex.api.entities.agents.OrganizationEntity;
-import com.lognex.api.entities.documents.DocumentEntity;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,23 +15,15 @@ import lombok.Setter;
 @NoArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class RetailDrawerCashOutEntity extends DocumentEntity {
-    private EmployeeEntity owner;
-    private Boolean shared;
     private AgentEntity agent;
-    private String externalCode;
+    private Boolean applicable;
+    private String created;
     private String description;
     private MetaEntity documents;
-    private String created;
-    private Boolean applicable;
-    private Long sum;
-    private Integer version;
+    private String externalCode;
     private String moment;
-    private String accountId;
-    private RateEntity rate;
     private OrganizationEntity organization;
-    private String name;
-    private String id;
-    private String updated;
+    private RateEntity rate;
     private RetailShiftEntity retailShift;
-    private GroupEntity group;
+    private Long sum;
 }
